[Bronze II] Title: 숫자, Time: 352 ms, Memory: 28136 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -2,7 +2,7 @@ import java.util.*;
 import java.io.*;
 
 public class Main {// 10093번 숫자 B2 23% 길래..
-	// 구현 뭐야 나도 틀렸네 방심하지 맙시당. 2분컷하느라 호
+	// 구현 뭐야 나도 틀렸네 방심하지 맙시당
 	static long A, B;
 	static long answer, cnt;
 	static StringBuilder sb;
@@ -12,7 +12,7 @@ public class Main {// 10093번 숫자 B2 23% 길래..
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine());
 		A = Long.parseLong(st.nextToken());
-		B = Integer.parseInt(st.nextToken());
+		B = Long.parseLong(st.nextToken());
 		long Big = Math.max(A, B);
 		long Small = Math.min(A, B);
